Introduce `with_attributes` view helper
Introduce the `with_attributes` view helper. Inspired by
`Object#with_options`, when the `with_attributes` helper is called with a block,
it yields a block argument that merges options into a base set of attributes.
For example:

```ruby
with_attributes class: "border rounded-sm p-4" do |styled|
  styled.link_to "I'm styled!", "/"
  # #=> <a class="border rounded-sm p-4" href="/">I'm styled!</a>
end
```

When the block is omitted, the object that would be the block parameter is
returned:

```ruby
 styled = with_attributes class: "border rounded-sm p-4"
 styled.link_to "I'm styled!", "/"
 # #=> <a class="border rounded-sm p-4" href="/">I'm styled!</a>
```

To change the receiver from the view context, pass an object as the first
argument:

```ruby
 button = with_attributes class: "border rounded-sm p-4"
 button.link_to "I have a border", "/"
 # #=> <a class="border rounded-sm p-4" href="/">I have a border</a>

 primary = with_attributes button, class: "text-red-500 border-red-500"
 primary.link_to "I have a red border", "/"
 # #=> <a class="border rounded-sm p-4 text-red-500 border-red-500" href="/">I have a red border</a>

 secondary = with_attributes button, class: "text-blue-500 border-blue-500"
 secondary.link_to "I have a blue border", "/"
 # #=> <a class="border rounded-sm p-4 text-blue-500 border-blue-500" href="/">I have a blue border</a>
```