Resolve `escape_attributes:` deprecation warning
In response to the following deprecation warning:

```
DEPRECATION WARNING: Use of the option :escape_attributes is
deprecated. It currently escapes both names and values of tags and
attributes and it is equivalent to :escape. If any of them are enabled,
the escaping is fully enabled.
```

Replace default `escape_attributes:` with `escape:`.